GHC.TcPluginM.Extra

Description

Create new constraints

newWanted :: CtLoc -> PredType -> TcPluginM CtEvidence

Create a new Wanted constraint with the given CtLoc.

newGiven :: EvBindsVar -> CtLoc -> PredType -> EvTerm -> TcPluginM CtEvidence

Create a new [G]iven constraint, with the supplied evidence. This must not be invoked from tcPluginInit or tcPluginStop, or it will panic.

Creating evidence

evByFiat

Arguments

:: String
Name the coercion should have

-> Type
The LHS of the equivalence relation (~)

-> Type
The RHS of the equivalence relation (~)

-> EvTerm

The EvTerm equivalent for unsafeCoerce

Lookup

lookupModule

Arguments

:: ModuleName
Name of the module

-> FastString
Name of the package containing the module. NOTE: This value is ignored on ghc>=8.0.

-> TcPluginM Module

Find a module

lookupName :: Module -> OccName -> TcPluginM Name

Find a Name in a Module given an OccName

Trace state of the plugin

tracePlugin :: String -> TcPlugin -> TcPlugin

Print out extra information about the initialisation, stop, and every run of the plugin when -ddump-tc-trace is enabled.

Substitutions

flattenGivens :: [Ct] -> [Ct]

Flattens evidence of constraints by substituting each others equalities.

NB: Should only be used on [G]iven constraints!

NB: Doesn't flatten under binders

mkSubst :: Ct -> Maybe ((TcTyVar, TcType), Ct)

Create simple substitution from type equalities

mkSubst' :: [Ct] -> [((TcTyVar, TcType), Ct)]

Create flattened substitutions from type equalities, i.e. the substitutions have been applied to each others right hand sides.

substType :: [(TcTyVar, TcType)] -> TcType -> TcType

Apply substitutions in Types

NB: Doesn't substitute under binders

substCt :: [(TcTyVar, TcType)] -> Ct -> Ct

Apply substitution in the evidence of Cts