GHC.TcPluginM.Extra

Description

Create new constraints

newWanted :: CtLoc -> PredType -> TcPluginM CtEvidence

Create a new [W]anted constraint.

newGiven :: CtLoc -> PredType -> EvTerm -> TcPluginM CtEvidence

Create a new [G]iven constraint, with the supplied evidence. This must not be invoked from tcPluginInit or tcPluginStop, or it will panic.

newDerived :: CtLoc -> PredType -> TcPluginM CtEvidence

Create a new [D]erived constraint.

Creating evidence

evByFiat

Arguments

:: String
Name the coercion should have

-> Type
The LHS of the equivalence relation (~)

-> Type
The RHS of the equivalence relation (~)

-> EvTerm

The EvTerm equivalent for unsafeCoerce

Lookup

lookupModule

Arguments

:: ModuleName
Name of the module

-> FastString
Name of the package containing the module

-> TcPluginM Module

Find a module

lookupName :: Module -> OccName -> TcPluginM Name

Find a Name in a Module given an OccName

Trace state of the plugin

tracePlugin :: String -> TcPlugin -> TcPlugin

Print out extra information about the initialisation, stop, and every run of the plugin when -ddump-tc-trace is enabled.